+++ packages/portal/stories/components/ManageableList/index.stories.tsx
@@ -1,9 +1,14 @@
 import * as React from 'react';
 import { DSProvider } from '@synerise/ds-core';
-import { storiesOf } from "@storybook/react";
-import { action } from "@storybook/addon-actions";
 import ManageableList from '@synerise/ds-manageable-list';
+import FileM from '@synerise/ds-icon/dist/icons/FileM';
+import Tag, { TagShape } from '@synerise/ds-tags/dist/Tag/Tag';
+import { storiesOf } from '@storybook/react';
 import { withState } from '@dump247/storybook-state';
+import { action } from '@storybook/addon-actions';
+import { ListType } from '@synerise/ds-manageable-list/dist/ManageableList';
+import { boolean } from '@storybook/addon-knobs';
+import FolderM from '@synerise/ds-icon/dist/icons/FolderM';
 
 const ITEMS:any = [
   {
@@ -12,6 +17,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: false,
     canDelete: false,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000001",
@@ -19,6 +25,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000002",
@@ -26,6 +33,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000003",
@@ -33,6 +41,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000004",
@@ -40,6 +49,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000005",
@@ -47,6 +57,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   },
   {
     id: "00000000-0000-0000-0000-000000000006",
@@ -54,6 +65,7 @@ const ITEMS:any = [
     canAdd: true,
     canUpdate: true,
     canDelete: true,
+    icon: <FolderM />
   }
 ];
 
@@ -63,7 +75,38 @@ const EMPTY_ITEM = {
   canAdd: true,
   canUpdate: true,
   canDelete: true,
-}
+  icon: <FolderM />
+};
+
+const CONTENT_ITEMS: any = [
+  {
+    id: "00000000-0000-0000-0000-000000000000",
+    name: "Position 0",
+    canAdd: true,
+    canUpdate: false,
+    canDelete: false,
+    tag: <Tag name={"A"} shape={TagShape.SINGLE_CHARACTER_ROUND} color={"red"} />,
+    content: <div>content</div>,
+  },
+  {
+    id: "00000000-0000-0000-0000-000000000002",
+    name: "Position 1",
+    canAdd: true,
+    canUpdate: true,
+    canDelete: true,
+    tag: <Tag name={"1"} shape={TagShape.SINGLE_CHARACTER_SQUARE} color={"#f3f5f6"} textColor={"#949ea6"} />,
+    content: <div>content</div>,
+  },
+  {
+    id: "00000000-0000-0000-0000-000000000001",
+    name: "Position 2",
+    canAdd: true,
+    canUpdate: true,
+    canDuplicate: true,
+    canDelete: true,
+    icon: <FileM />,
+  }
+];
 
 storiesOf('Components|Manageable List', module)
   .add('default', withState({
@@ -107,6 +150,8 @@ storiesOf('Components|Manageable List', module)
               addItemLabel="Add folder"
               showMoreLabel="show all"
               showLessLabel="show less"
+              more="more"
+              less="less"
               maxToShowItems={5}
               onItemAdd={addItem}
               onItemRemove={removeItem}
@@ -128,6 +173,8 @@ storiesOf('Components|Manageable List', module)
             addItemLabel="Add folder"
             showMoreLabel="show all"
             showLessLabel="show less"
+            more="more"
+            less="less"
             maxToShowItems={5}
             onItemAdd={action('onItemAdd')}
             onItemRemove={action('onItemRemove')}
@@ -139,4 +186,40 @@ storiesOf('Components|Manageable List', module)
         </div>
       </DSProvider>
     </div>
-  ));
+  ))
+  .add('content list', withState({
+    items: CONTENT_ITEMS,
+  })(({store}) => {
+    const handleChangeOrder = (newOrder) => {
+      store.set({items: newOrder});
+    };
+
+    return (
+      <div style={{width: '600'}}>
+        <DSProvider code="en_GB">
+          <div style={{ background: "#fff", width: '600px' }}>
+            <ManageableList
+              addItemLabel="Add position"
+              showMoreLabel="show all"
+              showLessLabel="show less"
+              more="more"
+              less="less"
+              maxToShowItems={5}
+              onItemAdd={action('onItemAdd')}
+              onItemRemove={action('onItemRemove')}
+              onItemEdit={action('onItemEdit')}
+              onItemSelect={action('onItemSelect')}
+              onItemDuplicate={action('onItemDuplicate')}
+              onChangeOrder={ boolean('Change order available', false) ? handleChangeOrder : null}
+              type={ListType.content}
+              items={store.state.items}
+              loading={false}
+              addButtonDisabled={boolean('Disable add item button', false)}
+              changeOrderDisabled={boolean('Disable change order', false)}
+              greyBackground={ boolean('Grey background', false) }
+            />
+          </div>
+        </DSProvider>
+      </div>
+    )
+  }));